add by hyb for db
